+++ lib/map.js
@@ -6,11 +6,11 @@ exports.parameter = (val) => {
   const pairs = val.split(',').map((pair) => pair.split('='));
   pairs.forEach((pair) => {
     if (pair[0] === 'ParameterKey') {
-      ret.parameterKey = pair[1];
+      ret.ParameterKey = pair[1];
     } else if (pair[0] === 'ParameterValue') {
-      ret.parameterValue = pair[1];
+      ret.ParameterValue = pair[1];
     } else if (pair[0] === 'UsePreviousValue') {
-      ret.usePreviousValue = pair[1] === 'true';
+      ret.UsePreviousValue = pair[1] === 'true';
     } else {
       throw new Error('ParameterKey, ParameterValue or UsePreviousValue expected');
     }
